requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
httpx>=0.24
pytest>=7.3
pytest-asyncio>=0.21
python-multipart>=0.0.6
pydantic>=2.0

# Para RAG con LangChain
langchain>=0.3,<0.4
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=1.0
langchain_openai>=0.1
langchain_core>=0.1
langchain_community>=0.1
gtts
